make disk formatting idempotent

+++ recipes/default.rb
@@ -37,6 +37,7 @@ else
 
     execute "Format Raid Array" do
       command "mkfs.#{node['ephemeral_mdadm']['filesystem']} #{node['ephemeral_mdadm']['raid_device']}"
+      not_if { `file -sL #{node['ephemeral_mdadm']['raid_device']}` =~ /filesystem/ }
     end
 
     directory node['ephemeral_mdadm']['mount_point'] do
